libs/zenn-ai-agent/pyproject.toml
[project]
name = "zenn-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "google-cloud-speech>=2.29.0",
    "google-cloud-texttospeech>=2.23.0",
    "google-genai>=0.3.0",
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "pyaudio>=0.2.14",
    "pydantic-settings>=2.7.0",
    "pyttsx3>=2.98",
    "speechrecognition>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
